fix: Linux build compatibility for aarch64 and modern toolchains
- Fix FFmpeg component names (AVCODEC/AVUTIL → avcodec/avutil) to match
  FindFFmpeg.cmake case-sensitive comparison
- Fix ENABLE_FRONTEND_API/ENABLE_QT to use option() instead of set() so
  CMake preset cache variables are respected
- Fix Qt6 alias target error with Qt 6.10+ by checking ALIASED_TARGET
  before setting properties
- Add Linux libssp linking case in ssp_connector/CMakeLists.txt
- Add Linux install targets for ssp-connector and libssp
- Skip ssp-connector build on aarch64 Linux (no vendor binary available)
- Fix int64_t vs long long type mismatch on aarch64 in obs-ssp-source.cpp

+++ src/obs-ssp-source.cpp
@@ -176,7 +176,7 @@ static void ssp_on_video_data(struct imf::SspH264Data *video, ssp_connection *s)
 		}
 	}
 
-	int64_t ts = video->pts;
+	long long ts = video->pts;
 	bool got_output;
 	bool success = ffmpeg_decode_video(&s->vdecoder, video->data,
 					   video->len, &ts, VIDEO_CS_DEFAULT,
